requirements.txt
yfinance==0.2.18
py_vollib==1.0.1
pandas==1.5.3
numpy==1.23.5
requests==2.28.1
python-dotenv==0.21.0
PyYAML==6.0
scipy==1.9.3
click==8.1.3